+++ index.ts
@@ -1,11 +1,16 @@
-import { curry } from "rambda";
-
-export default function decora<F, Alter extends (fn: F) => any>(fn: Alter) {
+/**
+ * Decorator factory that applies a function to a method.
+ * please enable "experimentalDecorators" in your tsconfig.json
+ */
+export default function decora<FnQ, FnR>(fn: (n: FnQ) => FnR) {
   return (
     target: any,
     key: PropertyKey,
-    descriptor: TypedPropertyDescriptor<F>
+    descriptor: TypedPropertyDescriptor<FnQ>,
   ) => {
-    descriptor.value = fn(descriptor.value!);
+    return {
+      ...descriptor,
+      value: fn(descriptor.value!),
+    } as TypedPropertyDescriptor<FnR>;
   };
 }